+++ app/src/main/java/org/jellyfin/androidtv/auth/model/User.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.androidtv.data.model
+package org.jellyfin.androidtv.auth.model
 
 import java.util.*
