--- app/src/main/java/org/jellyfin/androidtv/data/model/UserConfiguration.kt
@@ -1,8 +1,0 @@
-package org.jellyfin.androidtv.data.model
-
-/**
- * UserConfiguration model to use locally in place of UserConfiguration model in ApiClient.
- */
-data class UserConfiguration(
-	val latestItemsExcludes: List<String> = emptyList()
-)
