+++ app/src/main/java/org/jellyfin/androidtv/util/apiclient/ModelExtensions.kt
@@ -1,7 +1,7 @@
 package org.jellyfin.androidtv.util.apiclient
 
-import org.jellyfin.androidtv.data.model.Server
-import org.jellyfin.androidtv.data.model.User
+import org.jellyfin.androidtv.auth.model.Server
+import org.jellyfin.androidtv.auth.model.User
 import org.jellyfin.androidtv.util.toUUID
 import org.jellyfin.apiclient.discovery.DiscoveryServerInfo
 import org.jellyfin.apiclient.model.dto.UserDto
